Partition Image Files into Movement Sequences (#581)
Changed the rotation detection script to partition a set of sequential image files around rotation/invalid movement
