+++ localization/sparse_mapping/theia_map.md
@@ -130,15 +130,15 @@ double-precision seconds since epoch), which provides for a rather
 unique name, as compared to using the image index in the bag without
 that option.
 
+Partition image files into sequences:
+    rosrun sparse_mapping partition_image_sequences image_directory_name config_path
+
+Partitions a sequentially ordered set of image files into valid, rotation, and invalid sequences. During bundle adjustment, it is useful to avoid adding pure rotation sequences initially as these cause errors for monocular systems. The resulting sequences can be individually bundle-adjusted and merged as described later, generally starting with the valid (non-rotation) sequences and optionally adding rotations at the end once enough matches exist in the map. See 'rosrun sparse_mapping partition_image_sequences -h' for more usage details, options, and instructions.
+
 Remove low movement images:
     rosrun sparse_mapping remove_low_movement_images image_directory_name
 
 This will delete subsequent images with low movement from that directory to improve mapping performance and accuracy. 
-
-Remove rotation-only movement images:
-    rosrun sparse_mapping remove_rotation_only_images image_directory_name config_path
-
-Removes rotation only image sequences and optionally saves different movement sequences to different subdirectories. See 'rosrun sparse_mapping remove_rotation_only_images -h' for more usage details, options, and instructions.
 
 These are non-reversible operations, so they should be invoked on a copy
 of the images.
@@ -146,7 +146,7 @@ of the images.
 If possible, the robot should have some translation motion (in addition to any rotation) when
 the data is acquired.
 
-Removing low movement and rotation only movement images helps the accuracy of bundle adjustment, which struggles to optimize camera poses with small or no translation changes.
+Removing low movement images and initially only adding valid movement images helps the accuracy of bundle adjustment, which struggles to optimize camera poses with small or no translation changes.
 
 Put the selected images in a list:
